requirements.txt
# Core
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.12.5

# HTML parsing (DocConnector)
beautifulsoup4>=4.12.0

# Jira connector
jira>=3.6.0

# Google Drive connector
google-adk>=1.21.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-generativeai>=0.7.0
litellm>=1.80.16
